pyproject.toml
[project]
name = "qwen-synthetic-augmentation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "diffusers",
    "transformers",
    "pillow",
    "torch",
    "gguf>=0.10.0",
    "accelerate",
    "torchvision>=0.23.0",
    "jupyter>=1.1.1",
    "protobuf>=6.32.1",
    "sentencepiece>=0.2.1",
    "numpy>=2.2.6",
    "pycocotools>=2.0.10",
    "matplotlib>=3.10.6",
]

[tool.uv.sources]
diffusers = { git = "https://github.com/huggingface/diffusers" }
transformers = { git = "https://github.com/huggingface/transformers" }
torch = [
    { index = "pytorch-gpu" },
]
torchvision = [
    { index = "pytorch-gpu" },
]

[[tool.uv.index]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu129"
explicit = true
